chore(observatory): remove st-observatory and dead-code registry module
