+++ pyproject.toml
@@ -21,7 +21,6 @@ st-validate-local = "standard_tooling.bin.validate_local:main"
 st-ensure-label = "standard_tooling.bin.ensure_label:main"
 st-list-project-repos = "standard_tooling.bin.list_project_repos:main"
 st-set-project-field = "standard_tooling.bin.set_project_field:main"
-st-observatory = "standard_tooling.bin.observatory:main"
 st-generate-commands = "standard_tooling.bin.generate_commands:main"
 st-docker-run = "standard_tooling.bin.docker_run:main"
 st-docker-test = "standard_tooling.bin.docker_test:main"
