+++ src/standard_tooling/lib/github.py
@@ -43,3 +43,24 @@ def merge(pr: str, *, strategy: str, delete_branch: bool = True) -> None:
     if delete_branch:
         cmd.append("--delete-branch")
     run(*cmd)
+
+
+def list_project_repos(owner: str, project: str) -> list[str]:
+    """Return sorted, unique repos linked to a GitHub Project."""
+    jq_filter = (
+        f".[] | select(.projectsV2.Nodes | length > 0) "
+        f"| select(.projectsV2.Nodes[].number == {project}) "
+        f"| .nameWithOwner"
+    )
+    output = read_output(
+        "repo",
+        "list",
+        owner,
+        "--json",
+        "nameWithOwner,projectsV2",
+        "--limit",
+        "100",
+        "--jq",
+        jq_filter,
+    )
+    return sorted({r for r in output.splitlines() if r})
